requirements.txt
Flask==2.3.3
cryptography==41.0.3
gunicorn==20.1.0
mysql-connector-python==8.0.33
requests==2.31.0
